+++ R/fd_chull_intersect.R
@@ -1,11 +1,12 @@
 # A wrapper around geometry::intersectn() to properly handle errors and specific
 # cases
+#' @importFrom geometry intersectn
 fd_chull_intersect <- function(traits1, traits2) {
 
   traits1 <- traits1[!duplicated(traits1),, drop = FALSE]
   traits2 <- traits2[!duplicated(traits2),, drop = FALSE]
 
-  if (ncol(traits1) == 1L & ncol(traits2) == 1L) {
+  if (ncol(traits1) == 1L && ncol(traits2) == 1L) {
 
     # Range of the overlap
     r3 <- c(max(traits1, traits2),
@@ -22,7 +23,7 @@ fd_chull_intersect <- function(traits1, traits2) {
     )
   }
 
-  if (nrow(traits1) <= ncol(traits1) | nrow(traits2) <= ncol(traits2)) {
+  if (nrow(traits1) <= ncol(traits1) || nrow(traits2) <= ncol(traits2)) {
     return(
       list(
         "hull" = seq_len(nrow(traits1) + nrow(traits2)),
@@ -32,5 +33,5 @@ fd_chull_intersect <- function(traits1, traits2) {
     )
   }
 
-  return(geometry::intersectn(traits1, traits2, options = "FA")[["ch"]])
+  return(intersectn(traits1, traits2, options = "FA")[["ch"]])
 }