+++ tests/testthat/test-fdiv.R
@@ -1,16 +1,17 @@
-# Preamble code
+# Preamble code ----------------------------------------------------------------
 data("traits_birds")
 traits_birds_sc <- scale(traits_birds)
 
-# Actual tests
+
+# Tests for valid inputs -------------------------------------------------------
+
 test_that("Functional Divergence output format", {
 
   fdiv <- expect_silent(fd_fdiv(traits_birds))
 
   expect_s3_class(fdiv, "data.frame")
-  expect_length(fdiv, 2)
-  expect_equal(nrow(fdiv), 1)
-  expect_equal(colnames(fdiv), c("site", "FDiv"))
+  expect_identical(dim(fdiv), c(1L, 2L))
+  expect_named(fdiv, c("site", "FDiv"))
 
   expect_equal(fdiv$FDiv, 0.7282172, tolerance = 1e-7)
 })
@@ -38,7 +39,7 @@ test_that("Functional Divergence works for site with no species", {
     fd_fdiv(traits_plants, site_sp_plants[10,, drop = FALSE])
   )
 
-  expect_equal(fdiv$FDiv[[1]], 0)
+  expect_identical(fdiv$FDiv[[1]], 0)
 })
 
 test_that("Functional Divergence works in 1D", {
@@ -62,9 +63,8 @@ test_that("Functional Divergence works with sparse matrices", {
   fdiv <- expect_silent(fd_fdiv(traits_birds, sparse_site_sp))
 
   expect_s3_class(fdiv, "data.frame")
-  expect_length(fdiv, 2)
-  expect_equal(nrow(fdiv), 1)
-  expect_equal(colnames(fdiv), c("site", "FDiv"))
+  expect_identical(dim(fdiv), c(1L, 2L))
+  expect_named(fdiv, c("site", "FDiv"))
 
   expect_equal(
     fd_fdiv(traits_birds, sparse_site_sp),
@@ -72,6 +72,20 @@ test_that("Functional Divergence works with sparse matrices", {
   )
 
 })
+
+test_that("Functional Divergence works with matrices without rownames", {
+
+  traits_birds_un <- traits_birds
+  rownames(traits_birds_un) <- NULL
+
+  expect_identical(
+    fd_fdiv(traits_birds_un),
+    fd_fdiv(traits_birds)
+  )
+
+})
+
+# Tests for invalid inputs -----------------------------------------------------
 
 test_that("Functional Divergence fails gracefully", {
 
@@ -86,6 +100,18 @@ test_that("Functional Divergence fails gracefully", {
     fd_fdiv(data.frame(a = 1, row.names = "sp1"), matrix(1)),
     paste0("No species in common found between trait dataset ",
            "and site-species matrix"),
+    fixed = TRUE
+  )
+
+  ## Categorical trait data
+  # Add non-continuous traits
+  traits_birds_cat <- as.data.frame(traits_birds_sc)
+  traits_birds_cat$cat_trait <- "a"
+
+  expect_error(
+    fd_fdiv(traits_birds_cat, site_sp_birds),
+    paste0("Non-continuous trait data found in input traits. ",
+           "Please provide only continuous trait data"),
     fixed = TRUE
   )
 })