+++ tests/testthat/test-feve.R
@@ -1,25 +1,25 @@
-# Preamble code
+# Preamble code ----------------------------------------------------------------
 data("traits_birds")
 simple_site_sp <- matrix(1, nrow = 1, ncol = nrow(traits_birds),
                         dimnames = list("s1", row.names(traits_birds)))
 
-# Actual tests
+
+# Tests for valid inputs -------------------------------------------------------
+
 test_that("Functional Evenness output format", {
 
   feve <- expect_silent(fd_feve(traits_birds, sp_com = simple_site_sp))
 
   expect_s3_class(feve, "data.frame")
-  expect_length(feve, 2)
-  expect_equal(nrow(feve), 1)
-  expect_equal(colnames(feve), c("site", "FEve"))
+  expect_identical(dim(feve), c(1L, 2L))
+  expect_named(feve, c("site", "FEve"))
 
 
   feve <- expect_silent(fd_feve(traits_birds))
 
   expect_s3_class(feve, "data.frame")
-  expect_length(feve, 2)
-  expect_equal(nrow(feve), 1)
-  expect_equal(colnames(feve), c("site", "FEve"))
+  expect_identical(dim(feve), c(1L, 2L))
+  expect_named(feve, c("site", "FEve"))
 
 })
 
@@ -39,7 +39,7 @@ test_that("Functional Evenness computation are in line with other packages", {
 
   abund_mat <- matrix(1, ncol = 3, dimnames = list("site1", letters[1:3]))
 
-  expect_equal(fd_feve(sp_com = abund_mat, dist_matrix = test_dissim)$FEve, 1)
+  expect_identical(fd_feve(sp_com = abund_mat, dist_matrix = test_dissim)$FEve, 1)
 })
 
 test_that("Functional Evenness works in 1D", {
@@ -69,10 +69,9 @@ test_that("Functional Evenness works on subset of site/species", {
 test_that("Functional Evenness edge cases", {
 
   # n_species = 2 < 3 so not possible to compute FEve
-  expect_equal(fd_feve(traits_birds[1:2,],
+  expect_identical(fd_feve(traits_birds[1:2,],
                        simple_site_sp[, 1:2, drop = FALSE])[["FEve"]],
-               NA_real_)
-
+                   NA_real_)
 
   # n_species = 0, so no definition of FEve
   data("traits_plants")
@@ -82,7 +81,7 @@ test_that("Functional Evenness edge cases", {
     fd_feve(traits_plants, site_sp_plants[10,, drop = FALSE])
   )
 
-  expect_equal(feve$FEve[[1]], NA_real_)
+  expect_identical(feve$FEve[[1]], NA_real_)
 })
 
 test_that("Functional Evenness works on sparse matrices", {
@@ -100,22 +99,30 @@ test_that("Functional Evenness works on sparse matrices", {
   # Only site-species matrix is sparse
   expect_silent(fd_feve(traits_birds, sparse_site_sp))
 
-  expect_equal(fd_feve(traits_birds, sparse_site_sp)$FEve, 0.3743341,
-               tolerance = 1e-6)
+  expect_equal(
+    fd_feve(traits_birds, sparse_site_sp),
+    fd_feve(traits_birds, site_sp)
+  )
 
   # Only distance matrix is sparse
   expect_silent(fd_feve(sp_com = site_sp, dist_matrix = sparse_dist_mat))
 
-  expect_equal(fd_feve(sp_com = site_sp, dist_matrix = sparse_dist_mat)$FEve,
-               0.3743341, tolerance = 1e-6)
+  expect_equal(
+    fd_feve(sp_com = site_sp, dist_matrix = sparse_dist_mat),
+    fd_feve(sp_com = site_sp, dist_matrix = dist(traits_birds))
+  )
 
   # Both site-species and distance matrix are sparse
   expect_silent(fd_feve(sp_com = sparse_site_sp, dist_matrix = sparse_dist_mat))
 
   expect_equal(
-    fd_feve(sp_com = sparse_site_sp, dist_matrix = sparse_dist_mat)$FEve,
-    0.3743341, tolerance = 1e-6)
+    fd_feve(sp_com = sparse_site_sp, dist_matrix = sparse_dist_mat),
+    fd_feve(sp_com = site_sp, dist_matrix = dist(traits_birds))
+  )
 })
+
+
+# Tests for invalid inputs -----------------------------------------------------
 
 test_that("Functional Evenness fails gracefully", {
   # No traits and no dissimilarity
@@ -137,6 +144,18 @@ test_that("Functional Evenness fails gracefully", {
     fd_feve(data.frame(a = 1, row.names = "sp1"), matrix(1)),
     paste0("No species in common found between trait dataset ",
            "and site-species matrix"),
+    fixed = TRUE
+  )
+
+  ## Categorical trait data
+  # Add non-continuous traits
+  traits_birds_cat <- as.data.frame(traits_birds)
+  traits_birds_cat$cat_trait <- "a"
+
+  expect_error(
+    fd_feve(traits_birds_cat, site_sp_birds),
+    paste0("Non-continuous trait data found in input traits. ",
+           "Please provide only continuous trait data"),
     fixed = TRUE
   )
 })